==> vdj_analysis/3_assemble_incomplete_TCRs_MIXCR.R <==
#!/usr/bin/env Rscript

# Script to assemble mAb sequences F1-F4 for a given set of clonotypes
# 1. subset clonotypes of interest
# 2. create FASTAs for the clonotypes of interest (each clone should have a heavy and light chain)
# 3. use MIXCR to align input (or 'SEQUENCE_VDJ') to the germline and extract all the frame work regions
#  -- alternatively use AssignGenes.py igblast with the param [--format airr] (first try MIXCR)
library(alakazam)
library(shazam)
library(tigger)
library(dplyr)
library(data.table)
library(scales)
library(igraph)
library(devtools)
library(ggplot2)
library(stringr)
# contains some MIXCR functions
working_dir <- "./sc_analysis/"
setwd(working_dir)
source("Resources/functions_sc_analysis.R")


# immcantation results of overlapping samples only
THREADS     <- 10  # For MIXCR 
MIXCR       <- "./programs/mixcr-3.0.3/mixcr"

### INPUT:
IMM_RESULTS <- "./tables/mtcr_table_ALL_TCR.v3.1.csv"
RESULTS_DIR <- "./tables/incomplete_TCRs/"
dir.create(RESULTS_DIR)
output_csv  <- "./tables/rerun_MIXCR_on_incomplete_TCRs.4-14-20.tsv"


### 1. Load Immcantation results
imm_results <- read.csv(IMM_RESULTS)

germline_fasta_dir <- "~/share/germlines/imgt/human/vdj/"

### 2. List all cells missing either an Alpha/Beta AA sequence
data_tcr_na <- imm_results[is.na(imm_results$CDR3_AA),]
incomplete_tcrs <- as.character(unique(data_tcr_na$igseq_barcode))


### 3. Subset cells of interest with incomplete TCRs
imm_results.target <- imm_results[imm_results$igseq_barcode %in% incomplete_tcrs,]
target_clonotypes  <- unique(as.character(imm_results.target$MERGED_clonotypeID_imm))
target_cells       <- unique(as.character(imm_results.target$igseq_barcode))

target_input_seqs        <- as.character(imm_results.target$input_seq)
names_target_input_seqs  <- paste(imm_results.target$igseq_barcode,imm_results.target$chain,sep = ":")
names(target_input_seqs) <- names_target_input_seqs

### Need to automate this for all clones of interest

### 4. Create Master table for all target clonotype
#      - contains all the frame work regions generated by MIXCR
all_results <- list()

for (clone in target_clonotypes) {
  # subset 10X results
  candidates.df.clone <- imm_results.target[imm_results.target$MERGED_clonotypeID_imm %in% clone,]
  if (nrow(candidates.df.clone) <= 1){
    cat(paste0(">> Skip Clonotype ",clone,": insufficient results (usually meaning 1 chain)"))
    next
  }
  
  # If there are more than 2 chains, remove cells that only have one chain
  candidates.df.clone <- candidates.df.clone[! candidates.df.clone$chain_count %in% c("1"),]
  if (nrow(candidates.df.clone) < 2){
    cat(paste0(">> Skip Clonotype ",clone,": No Alpha and Beta chain pair for this clonotype"))
    next
  }
  
  # number of cells
  cells_vec <- unique(as.character(candidates.df.clone$igseq_barcode))
  
  # Format FASTA entry for a given clonotype (Heavy and Light chain)
  for (cell in cells_vec) {
    candidates.df.clone.cell <- candidates.df.clone[candidates.df.clone$igseq_barcode == cell,]
    fasta_id_vec  <- c()
    fasta_seq_vec <- c()
    for (i in 1:nrow(candidates.df.clone.cell)) {
      df.row <- candidates.df.clone.cell[i,]
      id_cell         <- as.character(df.row$igseq_barcode)
      id_sample       <- as.character(df.row$sample_igseq)
      id_clonotype    <- as.character(df.row$MERGED_clonotypeID_imm)
      id_chain        <- as.character(df.row$chain)
      fasta_id  <- paste(c(id_clonotype,id_cell,id_chain),collapse = ":")
      
      # Get Nucleotide sequence from immcantation results
      fasta_seq <- as.character(df.row$input_seq) 
      
      fasta_id_vec  <- c(fasta_id_vec,fasta_id)
      fasta_seq_vec <- c(fasta_seq_vec,fasta_seq)
    }
    
    fasta_entry        <- fasta_seq_vec
    names(fasta_entry) <- fasta_id_vec
    
    # Separate Heavy and Light chains
    chains_vec  <- tstrsplit(fasta_id_vec,":")[[3]]
    beta_chain  <- chains_vec[grep("TRB",chains_vec)]
    alpha_chain <- chains_vec[grep("TRA",chains_vec)]
    
    fasta_entry_beta  <- fasta_entry[grep("TRB",chains_vec)]
    fasta_entry_alpha <- fasta_entry[grep("TRA",chains_vec)]
    
    # Output FASTA entries
    ### Create clonotype output directory
    cluster_output_dir <- paste(c(RESULTS_DIR,clone,':',cell,'/'),collapse = "")
    dir.create(cluster_output_dir)
    
    # write Beta chain FASTA
    cluster_fasta_beta_fh <- paste(c(cluster_output_dir,clone,'.beta.fasta'),collapse = "")
    writeFasta(fasta_entry_beta,file =  cluster_fasta_beta_fh, width = 300)
    # write Alpha chain FASTA
    cluster_fasta_alpha_fh <- paste(c(cluster_output_dir,clone,'.alpha.fasta'),collapse = "")
    writeFasta(fasta_entry_alpha,file =  cluster_fasta_alpha_fh, width = 300)
    
    # Run MIXCR Heavy & Light
    results_list        <- list()
    results_beta       <- runMixcrPerChain(fasta = cluster_fasta_beta_fh,results_dir = cluster_output_dir,chain = "beta",threads = THREADS,MIXCR = MIXCR)
    results_beta_fr    <- results_beta[,names(results_beta)[grep("nSeqFR",names(results_beta))]]
    results_beta_cdr   <- results_beta[,names(results_beta)[grep("SeqCDR",names(results_beta))]]
    results_beta_score <- results_beta[,names(results_beta)[grep("WithScore",names(results_beta))]]
    results_beta_sub   <- results_beta[,c(names(results_beta_score),names(results_beta_fr),names(results_beta_cdr))]
    results_list[[beta_chain]] <- results_beta_sub
    ## DEBUG:
    cat(paste0(">>> BETA: ", clone, "\t", cell, "\n"))
    
    results_alpha       <- runMixcrPerChain(fasta = cluster_fasta_alpha_fh,results_dir = cluster_output_dir,chain = "alpha",threads = THREADS,MIXCR = MIXCR)
    results_alpha_fr    <- results_alpha[,names(results_alpha)[grep("nSeqFR",names(results_alpha))]]
    results_alpha_cdr   <- results_alpha[,names(results_alpha)[grep("SeqCDR",names(results_alpha))]]
    results_alpha_score <- results_alpha[,names(results_alpha)[grep("WithScore",names(results_alpha))]]
    results_alpha_sub   <- results_alpha[,c(names(results_alpha_score),names(results_alpha_fr),names(results_alpha_cdr))]
    results_list[[alpha_chain]] <- results_alpha_sub
    ## DEBUG:
    cat(paste0(">>> ALPHA: ", clone, "\t", cell, "\n"))
    
    # Merge MIXCR results back into candidate table
    master_cols     <- c(names(candidates.df.clone),names(results_beta_sub))
    merged_clone_df <- as.data.frame(matrix(nrow = 0,ncol = length(master_cols)))
    names(merged_clone_df) <- master_cols
    
    # If either Alpha or Beta has no MIXCR results, do not include cell in final results
    for (i in 1:nrow(candidates.df.clone.cell)) {
      orig_row        <- candidates.df.clone.cell[i,]
      chain           <- as.character(orig_row$chain)
      results_mixcr   <- results_list[[chain]]
      if (nrow(results_mixcr) == 0){
        next
      }
      combined_row    <- cbind(orig_row,results_mixcr)
      merged_clone_df <- rbind(merged_clone_df,combined_row)
    }
    
    # SKIP if MIXCR fails to assemble one chain 
    if (nrow(merged_clone_df) < 2){
      next
    }
    
    # Merge with master table spanning all target clonotypes
    entry <- paste(c(clone,cell),collapse = ":")
    all_results[[entry]] <- merged_clone_df
  }
  
}

cat(paste0("\n>>> COMBINE ALL MIXCR RESULTS\n"))
merged_results <- do.call(rbind,all_results)
row.names(merged_results) <- NULL


### OUTPUT - FINAL TABLE TSV (CSV is too annoying to deal with)
write.table(merged_results,file = output_csv,quote = F,row.names = F,sep = "\t")
